controller/requirements.txt
# apt install python3-dev
paho-mqtt
lirc
RPi.GPIO
pyserial